The VBA for ttts (An Excel User Defined Function)
Transfer To TiddlySpace TTTS V0.1
Usage: put the following in a cell at the end of each row of data that describes the tiddler(s) you want created/ updated
=ttts("user:password", "PUT",<cellContainingBodyofTiddler>, "application/json",<urlroot>&<tiddlerName>,<tag>)
'<cellContainingBodyofTiddler> should be a reference to a cell containing the body of the tiddler
<urlroot> should be a reference to a cell containing the baseurl for all tiddlers in the space
eg http://<space>.tiddlyspace.com:80/bags/<space>_public/tiddlers/
<tiddlerName> should be a ref to the cell containing the name of the tiddler to be created
<tag> should be a ref to the tag. Only a single tag is supported at present

The purpose of this user-defined function is to facilitate migration/ transfer of data from excel to tiddlyspaace
The reverse function is a simpler use case and can easily be implemented if required.
Limitations:
No warranty or guarantees with this.  Only use this if you understand what it is doing.
There is no collision detection & minimal error handling
Success recorded as OK + the time.
When called the function over-writes the destination tiddler with the data from excel.
So Excel acts as the master.  Any changes made to the excel spreadsheet are immediately replicated to tiddlyspace
No provision for configuring a proxy server as yet
Base address is embedded on First page of spreadsheet cell $E1
Only one Tag is currently supported by the json wrapper.

James Savage July 2012